renamed generic post serializer to abstract serializers
