+++ api/src/common/abstract_serializers.py
@@ -2,12 +2,12 @@ from django.db.models.functions import Length
 from comments.serializers import CommentDetailSerializer
 from users.serializers import BasicUserSerializer
 from rest_framework import serializers
-from .generic_serializers import GenericSerializer
+from common.generics.generic_serializers import GenericSerializer
 from photos.serializers import PhotoRetrieveSerializer
 
 
-# Serializes a generic resource model
-class GenericPostSerializer(GenericSerializer):
+# Serializes an abstract resource model
+class AbstractModelSerializer(GenericSerializer):
     creator = BasicUserSerializer(read_only=True)
     liked_users = BasicUserSerializer(read_only=True, many=True)
     is_liked = serializers.SerializerMethodField()
@@ -42,7 +42,7 @@ class GenericPostSerializer(GenericSerializer):
         return GenericSerializer.create(self, validated_data)
 
 
-# Serializes a generic resource model with detail
-class GenericPostDetailSerializer(GenericPostSerializer):
+# Serializes an abstract resource model with detail
+class AbstractModelDetailSerializer(AbstractModelSerializer):
     tagged_users = BasicUserSerializer(read_only=True, many=True)
     photos = PhotoRetrieveSerializer(read_only=True, many=True)
