+++ api/src/listings/serializers.py
@@ -1,9 +1,9 @@
-from common.generics.generic_post_serializers import GenericPostSerializer, GenericPostDetailSerializer
+from common.abstract_serializers import AbstractModelSerializer, AbstractModelDetailSerializer
 from .models import Listing
 
 
 # Serializes the listing model
-class ListingSerializer(GenericPostSerializer):
+class ListingSerializer(AbstractModelSerializer):
 
     class Meta:
         model = Listing
@@ -12,5 +12,5 @@ class ListingSerializer(GenericPostSerializer):
 
 
 # Serializes the listing model with detail
-class ListingDetailSerializer(ListingSerializer, GenericPostDetailSerializer):
+class ListingDetailSerializer(ListingSerializer, AbstractModelDetailSerializer):
     pass
